pyproject.toml
[project]
name = "image-gallery-auth-api"
version = "0.1.0"
description = "FastAPI project with JWT auth, MySQL, image gallery upload, drag-drop UI, and OCR"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "pymysql>=1.1.0",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "pydantic-settings>=2.4.0",
  "pillow>=10.4.0",
  "pytesseract>=0.3.10",
  "jinja2>=3.1.4",
  "email-validator>=2.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
